third_party/WebKit/LayoutTests/inspector/console/console-log-side-effects-expected.txt

CONSOLE MESSAGE: line 14: string
CONSOLE MESSAGE: line 15: 42
CONSOLE MESSAGE: line 16: false
CONSOLE MESSAGE: line 17: undefined
CONSOLE MESSAGE: line 18: null
CONSOLE MESSAGE: line 19: NaN
CONSOLE MESSAGE: line 20: -Infinity
CONSOLE MESSAGE: line 21: 0
CONSOLE MESSAGE: line 22: 42.0000
CONSOLE MESSAGE: line 23: -4.24200e-11
CONSOLE MESSAGE: line 24: true
CONSOLE MESSAGE: line 25: foo
CONSOLE MESSAGE: line 26: [object Object]
CONSOLE MESSAGE: line 27: [object Window]
CONSOLE MESSAGE: line 33: [object HTMLParagraphElement]
CONSOLE MESSAGE: line 37: [object Object]
CONSOLE MESSAGE: line 38: 1,2,3
CONSOLE MESSAGE: line 40: [object Uint32Array]
CONSOLE MESSAGE: line 43: 1,2,3,[object Object]
CONSOLE MESSAGE: line 46: [object Object]
CONSOLE MESSAGE: line 49: 1,2,3,[object Object],[object Object]
CONSOLE MESSAGE: line 60: 1,2,3,4,5,,,,,,NaN,,,,,,,NaN,1,2,3,4,5,,
CONSOLE MESSAGE: line 68: [object Object]
CONSOLE MESSAGE: line 71: 1,2,3,[object Object],[object Object],[object Object]
CONSOLE MESSAGE: line 77: 
CONSOLE MESSAGE: line 77: 
CONSOLE MESSAGE: line 77: 
CONSOLE MESSAGE: line 77: 
CONSOLE MESSAGE: line 77: 
CONSOLE MESSAGE: line 80: 
CONSOLE MESSAGE: line 86: 
CONSOLE MESSAGE: line 91: 
Tests various extreme usages of console.log()
